+++ doc/CHANGES.rst
@@ -17,6 +17,7 @@ CHANGELOG
 ------
 
 * Fix phone number with an empty string value
+* Fix migration with added more than one new required columns with default values
 
 0.19.1 (2018-06-07)
 -------------------